Add docstring to `typestree` (#219)
Added a docstring to `typestree`, because I could not find the function when I went looking for it, despite knowing that it existed. 

The only reference I could find is https://github.com/JuliaLang/julia/issues/24741, which I read some time ago, and is the only reason I know that it existed.

This PR should make it so that when you search for "supertype", "subtypes", or "tree", the function should appear in the documentation, which was the original problem.