Improve transition rules (#223)
This PR adds a much more complicated example for defining transition rules, as well as significantly changing the wording.

The old documentation was not enough for me to understand how to define the rules, so this took some trial and error. I believe that the documentation in the PR would have made things more clear from the start.