Remove 1.0 warning (#225)
With 2.0 out, it seemed appropriate to remove a warning about breaking changes made in 1.0, especially since it was so prominent. I may be wrong on that, I have no idea what the timelines are and what amount of grief you get from users about the changes informaed about in the warning.